pyproject.toml
[project]
name = "ai-blog-app"
version = "1.0.4"
description = "A modular AI application for creating high-engagement blog posts. This framework supports multiple LLMs, optional PDF context, and customizable article length."
readme = "README.md"
authors = [
    { name = "Edilma Riano", email = "EdilmaDev@Gmail.com" }
]
license = {  file = "LICENSE" }
requires-python = ">=3.10"
keywords = ["ai", "blog", "autogen", "llm", "agent", "content-generation"]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries",
    "Topic :: Text Processing :: General",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
    "Topic :: Utilities",
]

dependencies = [
    "autogen-agentchat>=0.7.1",
    "autogen-ext[all]>=0.7.1",
    "openai>=1.98.0"
]

[project.urls]
Homepage = "https://github.com/edilma/AI_Blog_App"
Documentation = "https://github.com/edilma/AI_Blog_App/blob/main/README.md"
BugTracker = "https://github.com/edilma/AI_Blog_App/issues"

[project.optional-dependencies]
fastapi = [
    "fastapi",
    "uvicorn[standard]",
    "python-dotenv"
]
dev = [
    "python-dotenv",
]

[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

# --- Package Discovery ---
[tool.setuptools.packages.find]
where = ["."]
include = ["ai_blog_app*"]

[tool.setuptools.package-data]
"ai_blog_app" = ["llm_config.json"]
